requirements.txt
numpy
onnxruntime
onnxruntime_gpu
Pillow
